refactor api routes
