+++ accounts/urls.py
@@ -1,14 +1,10 @@
 from django.conf.urls import patterns, url, include
 from rest_framework import routers
-from accounts.user_api import views
-
-router = routers.DefaultRouter()
-router.register(r'user', views.UserView, 'user')
+from django.contrib import admin
  
 admin.autodiscover()
 
 urlpatterns = patterns('',
-    url(r'^', include(router.urls)),
     url(r'^$', 'accounts.views.home'),
     url(r'^login/$', 'accounts.views.login',name='login'),
     url(r'^logout/$', 'accounts.views.logout',name='logout'),
